add SetDefaultPermissions to apisocket package
